Full coverage!!!
Adapter             100% files, 100% lines
Annotation       100% files, 100% lines
Bus                    100% files, 100% lines
Command         100% files, 100% lines
Configuration   100% files, 100% lines
Event                100% files, 100% lines
Gate                  100% files, 100% lines
Gate.php           100% lines
Message.php    100% lines

+++ src/Cqrs/Gate.php
@@ -43,11 +43,12 @@ class Gate {
     public function reset()
     {
         foreach($this->buses as $bus){
+            if($bus->getName()==='system-bus'){
+                continue;
+            }
             $this->detach($bus);
         }
-        $this->buses = array();
-        //return self::getInstance();
-        return $this; //self::getInstance();
+        return $this;
     }
 
     /**
@@ -60,11 +61,7 @@ class Gate {
                 new ClassMapCommandHandlerLoader(),
                 new ClassMapEventListenerLoader()
             );
-            try {
-                $this->attach($systemBus);
-            } catch ( GateException $e ){
-                echo $e->getMessage();
-            }
+            $this->attach($systemBus);
         }
     }
